pyproject.toml
[tool.poetry]
name = "well"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
fastapi = "0.74.0"
uvicorn = {extras = ["standard"], version = "0.17.5"}
SQLAlchemy = "1.4.31"
alembic = "1.7.6"
psycopg2-binary = "2.9.3"
pytest = "7.0.1"
factory-boy = "3.2.1"
pytest-factoryboy = "2.1.0"
SQLAlchemy-Utils = "0.38.2"
requests = "2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
black = "22.1.0"
isort = "5.10.1"
rope = "0.22.0"
flake8-isort = "4.1.1"
mypy = "0.931"
sqlalchemy-stubs = "0.4"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
